# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Restful Provisioning Protocol (RPP) ",
    "description": "A Modern Domain Registry API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/domains/": {
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Create Domain",
        "description": "Create a new domain",
        "operationId": "do_create_domains__post",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/{domainname}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get Domain Info",
        "operationId": "do_info_domains__domainname__get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Filter"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Delete Domain",
        "operationId": "do_delete_domains__domainname__delete",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Domains"
        ],
        "summary": "Update Domain",
        "operationId": "do_update_domains__domainname__patch",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/{domainname}/availability": {
      "head": {
        "tags": [
          "Domains"
        ],
        "summary": "Check Domain Availability",
        "operationId": "do_check_head_domains__domainname__availability_head",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Check Domain Availability",
        "operationId": "do_check_get_domains__domainname__availability_get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/{domainname}/processes/renewals": {
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Renew Domain",
        "operationId": "do_renew_domains__domainname__processes_renewals_post",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainRenewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/domains/{domainname}/processes/transfers/latest": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Query Transfer Status",
        "operationId": "do_query_transfer_domains__domainname__processes_transfers_latest_get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "domainname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domainname"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DomainCreateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PeriodModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "ns": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NsItemModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ns",
            "default": []
          },
          "registrant": {
            "type": "string",
            "title": "Registrant"
          },
          "contact": {
            "items": {
              "$ref": "#/components/schemas/ContactModel"
            },
            "type": "array",
            "title": "Contact"
          },
          "authInfo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthInfoModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "dnssec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DsOrKeyType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "registrant",
          "contact"
        ],
        "title": "DomainCreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}